Extract child management logic from ParentNode
